@@ -1,41 +1,38 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-        <Company>$(Authors)</Company>
-        <Authors>stagecodes</Authors>
-        <Description>Simple memory reader and writer for 32 and 64-bit processes</Description>
-        <Copyright>Copyright (c) StageCodes 2022. Licensed under the GNU General Public License v3.</Copyright>
-        <PackageProjectUrl>https://www.nuget.org/packages/SimpleMem/</PackageProjectUrl>
-        <PackageIcon>chip128.png</PackageIcon>
-        <PackageReadmeFile>README.md</PackageReadmeFile>
-        <RepositoryUrl>https://github.com/hburn7/SimpleMem</RepositoryUrl>
-        <AssemblyVersion>1.1.0</AssemblyVersion>
-        <PackageTags>memory,mem,memory read,memory write,read,write,hex,hexadecimal,pointer,pointer scan,pointerscan,offset,offsets,hack,hacking</PackageTags>
-        <RepositoryType>git</RepositoryType>
-        <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
-        <Version>1.1.0</Version>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-        <PackageReleaseNotes>Added 64-bit support, generic read/writes, AoB scan support</PackageReleaseNotes>
-        <PackageVersion>1.1.1</PackageVersion>
-        <PackageLicenseUrl></PackageLicenseUrl>
-        <PackageIconUrl>..\..\..\artwork\simple-mem\chip128.png</PackageIconUrl>
-        <Title>SimpleMem</Title>
+      <TargetFramework>net8.0</TargetFramework>
+      <ImplicitUsings>disable</ImplicitUsings>
+      <Nullable>disable</Nullable>
+      <GeneratePackageOnBuild Condition="'$(Configuration)' == 'Release'">True</GeneratePackageOnBuild>
+      <Company>$(Authors)</Company>
+      <Authors>stagecodes</Authors>
+      <Description>Simple memory reader and writer for 32 and 64-bit processes</Description>
+      <Copyright>Copyright (c) StageCodes 2022. Licensed under the GNU General Public License v3.</Copyright>
+      <PackageProjectUrl>https://github.com/spreedated/SimpleMem</PackageProjectUrl>
+      <PackageIcon>logo.png</PackageIcon>
+      <PackageReadmeFile>README.md</PackageReadmeFile>
+      <RepositoryUrl>https://github.com/spreedated/SimpleMem</RepositoryUrl>
+      <AssemblyVersion>1.2.0</AssemblyVersion>
+      <PackageTags>memory,mem,memory read,memory write,read,write,hex,hexadecimal,pointer,pointer scan,pointerscan,offset,offsets,hack,hacking</PackageTags>
+      <RepositoryType>git</RepositoryType>
+      <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
+      <Version>1.2.0</Version>
+      <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+      <PackageReleaseNotes>Added 64-bit support, generic read/writes, AoB scan support</PackageReleaseNotes>
+      <PackageVersion>1.2.0</PackageVersion>
+      <PackageLicenseUrl></PackageLicenseUrl>
+      <PackageIconUrl>..\logo.png</PackageIconUrl>
+      <Title>SimpleMem</Title>
+      <LangVersion>latest</LangVersion>
     </PropertyGroup>
 
-    <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-        <DocumentationFile>bin\Debug\SimpleMem.xml</DocumentationFile>
-    </PropertyGroup>
-
-    <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
+    <PropertyGroup Condition="'$(Configuration)' == 'Release'">
         <DocumentationFile>bin\Release\SimpleMem.xml</DocumentationFile>
     </PropertyGroup>
 
     <ItemGroup>
-        <None Include="..\..\..\artwork\simple-mem\chip128.png">
+        <None Include="..\logo.png">
             <Pack>True</Pack>
             <PackagePath>\</PackagePath>
         </None>
